gfx: a coordinate domain of 4096, and the segment steps in sixteen bits
The segment was the last place computing wider than the rest, and the
question of what it would take to narrow it turned out to have a
measured answer rather than an algebraic one.

The error term is tested at twice its value, and it reaches THREE TIMES
the segment's span - not once, as the obvious bound suggests. So the
coordinate range that keeps the stepping inside sixteen bits is far
tighter than it looks: a span of ten thousand, not sixty. Instrumenting
the exact stepping over the corners and slopes of squares of several
sizes is what produced that, and it corrected two estimates on the way.

4096 is the domain, and it is chosen against what the parts actually
drive: a panel a microcontroller runs is at most some hundreds of pixels
a side, so this leaves room to place a shape several screens outside one
and let the clipping deal with it, while the widest intermediate in the
stepping stays at 24574 - a quarter of the type still spare. The
classic producer of a distant coordinate is not a large panel but a long
list scrolled by a pixel offset, and that is what the headroom is for.

Declared where it can be ENFORCED rather than merely stated: a surface's
dimensions are template parameters, so a surface larger than the domain
fails to compile. A coordinate a caller computes cannot be checked
without paying for the question on every call, which is the reason the
bound is generous instead of tight.

What it bought, on an AVR: the segment routine from 742 bytes of code to
518. Together with the clip, a program drawing frames and lines goes
from 980 bytes to 818, and a complete graphical program with text from
4082 to 3970. The CH32V00x pays nothing for any of it - the locals are
fast types, so a 32-bit part keeps its native word.

The family fixture now exercises the domain's edges instead of standing
outside them: it had been drawing a segment between coordinates twenty
thousand apart, which under this contract is not a legal call.

clip() itself stays correct BEYOND the domain, since it never forms the
sum that would need the width. That is robustness and not a promise, and
the tests that push it to thirty-two thousand are kept as such.

Verification: test_gfx 169232 assertions over 38 cases, every segment on
a small surface still matching the reference exactly; ctest 27/27; brio
check all OK over the six strata; brio prose clean.

+++ brio/gfx/draw.hpp
@@ -148,22 +148,25 @@ void line(S& s, Coord x0, Coord y0, Coord x1, Coord y1, typename S::Color c) {
         return;
     }
 
-    int32_t x = x0;
-    int32_t y = y0;
-    const int32_t dx =
-        x1 > x0 ? static_cast<int32_t>(x1) - x0 : static_cast<int32_t>(x0) - x1;
-    const int32_t dy =
-        -(y1 > y0 ? static_cast<int32_t>(y1) - y0 : static_cast<int32_t>(y0) - y1);
-    const int32_t sx = x0 < x1 ? 1 : -1;
-    const int32_t sy = y0 < y1 ? 1 : -1;
-    int32_t err = dx + dy;
+    // Sixteen bits is enough for all of this INSIDE THE COORDINATE
+    // DOMAIN, and only there: the error is tested at twice its value and
+    // reaches three times the segment's span, so at the domain's edge
+    // the widest intermediate here is 24574. gfx/surface.hpp's
+    // `coord_max` is what buys that, and a fast type is what keeps a
+    // 32-bit part from paying to narrow.
+    int_fast16_t x = x0;
+    int_fast16_t y = y0;
+    const int_fast16_t dx = x1 > x0 ? x1 - x0 : x0 - x1;
+    const int_fast16_t dy = -(y1 > y0 ? y1 - y0 : y0 - y1);
+    const int_fast16_t sx = x0 < x1 ? 1 : -1;
+    const int_fast16_t sy = y0 < y1 ? 1 : -1;
+    int_fast16_t err = dx + dy;
     for (;;) {
-        s.fill_rect(static_cast<Coord>(x),
-                static_cast<Coord>(y), 1, 1, c);
-        if (x == static_cast<int32_t>(x1) && y == static_cast<int32_t>(y1)) {
+        s.fill_rect(static_cast<Coord>(x), static_cast<Coord>(y), 1, 1, c);
+        if (x == x1 && y == y1) {
             return;
         }
-        const int32_t e2 = 2 * err;
+        const int_fast16_t e2 = 2 * err;
         if (e2 >= dy) {
             err += dy;
             x += sx;
